[#159] TableListView, TableListViewCell 컴포넌트 코드 정리
