+++ MC2-OneShot/Presentation/Component/TableListView.swift
@@ -8,11 +8,12 @@
 import SwiftUI
 import SwiftData
 
+// MARK: - TableListView
+
 struct TableListView: View {
     
-    @Environment(PartyUseCase.self) var partyUseCase
+    @Environment(PartyUseCase.self) private var partyUseCase
     @Environment(HomePathModel.self) private var homePathModel
-    @Environment(\.modelContext) private var modelContext
     
     @State private var isShowAlert = false
     @State private var selectedParty: Party?
@@ -33,7 +34,7 @@ struct TableListView: View {
             .swipeActions {
                 Button {
                     selectedParty = party
-                    self.isShowAlert.toggle()
+                    isShowAlert.toggle()
                 } label: {
                     Text("삭제하기")
                 }
@@ -45,8 +46,8 @@ struct TableListView: View {
                 } else {
                     Button("지우기", role: .destructive) {
                         guard let selectedParty = selectedParty else { return }
+                        partyUseCase.deleteParty(selectedParty)
                         HapticManager.shared.notification(type: .success)
-                        modelContext.delete(selectedParty)
                     }
                     
                     Button("살리기", role: .cancel) {}
@@ -58,91 +59,9 @@ struct TableListView: View {
     
     /// 리스트에서 파티를 삭제할 때 출력되는 텍스트입니다.
     private var deletePartyAlertText: String {
-        guard let selectedParty = selectedParty else { return "" }
+        guard let selectedParty = selectedParty else { return "선택된 파티가 없습니다!" }
         if selectedParty.isLive { return "진행중인 술자리는 지울 수 없어,,," }
         else { return "진짤루?\n 술자리 기억...지우..는거야..?" }
-    }
-}
-
-// MARK: - TableListCellView
-struct TableListCellView: View {
-    
-    let thumbnail: UIImage
-    let title: String
-    let captureDate: Date
-    let isLive: Bool
-    let stepCount: Int
-    let notiCycle: Int
-    
-    var body: some View {
-        HStack {
-            Image(uiImage: thumbnail)
-                .resizable()
-                .frame(width: 68, height: 68)
-                .clipShape(RoundedRectangle(cornerRadius: 7.5))
-            
-            Spacer()
-                .frame(width: 12)
-            
-            VStack(alignment: .leading, spacing: 4) {
-                Text(title)
-                    .pretendard(.bold, 17)
-                    .foregroundStyle(.shotFF)
-                
-                Text("\(captureDate.yearMonthDay)")
-                    .pretendard(.regular, 14)
-                    .foregroundStyle(.shot6D)
-            }
-            
-            Spacer()
-            
-            VStack(spacing: 4) {
-                TableListStateInfoLabel(
-                    stateInfo: isLive ? .live : .end,
-                    stepCount: stepCount
-                )
-                
-                Text("\(notiCycle)min")
-                    .pretendard(.regular, 14)
-                    .foregroundStyle(.shot6D)
-            }
-        }
-        .background(.shot00)
-    }
-}
-
-// MARK: - TableListStateInfoLabel
-private struct TableListStateInfoLabel: View {
-    
-    /// 술자리 상태를 나타내는 열거형
-    enum StateInfo: String {
-        case live
-        case end
-    }
-    
-    let stateInfo: StateInfo
-    let stepCount: Int
-    
-    var body: some View {
-        if stateInfo == .live {
-            Text("LIVE")
-                .pretendard(.bold, 14)
-                .foregroundStyle(Color.shot00)
-                .frame(width: 76, height: 22)
-                .background(Color.shotGreen)
-                .clipShape(RoundedRectangle(cornerRadius: 12))
-        } else {
-            HStack(spacing: 0) {
-                Text("STEP ")
-                    .foregroundStyle(Color.shotD8)
-                Text("\(stepCount.intformatter)")
-                    .foregroundStyle(Color.shotGreen)
-            }
-            .pretendard(.bold, 14)
-            .frame(width: 76, height: 22)
-            .background(Color.shot33)
-            .clipShape(RoundedRectangle(cornerRadius: 12))
-        }
     }
 }
 
@@ -154,16 +73,5 @@ private struct TableListStateInfoLabel: View {
     return TableListView()
         .environment(HomePathModel())
         .modelContainer(modelContainer)
-}
-
-#Preview {
-    TableListCellView(
-        thumbnail: UIImage(resource: .test),
-        title: "테스트 제목",
-        captureDate: .now,
-        isLive: true,
-        stepCount: 7,
-        notiCycle: 30
-    )
 }
 #endif
